back-end/requirements.txt
Flask
flask
flask-cors
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-bcrypt
requests
werkzeug
